new updates added with reset,forgot,email verify etc

+++ controllers/orderController.js
@@ -1,6 +1,16 @@
 const Order = require("../models/orderModel");
 const Cart = require("../models/cartModel");
 const Product = require("../models/productModel");
+const Razorpay = require("razorpay");
+const crypto = require("crypto");
+const dotenv = require("dotenv");
+
+dotenv.config();
+
+const instance = new Razorpay({
+  key_id: process.env.RAZORPAY_KEY,
+  key_secret: process.env.RAZORPAY_SECRET,
+});
 
 exports.getAllOrders = async (req, res) => {
   try {
@@ -20,26 +30,48 @@ exports.getAllOrders = async (req, res) => {
 exports.createOrder = async (req, res) => {
   try {
     //  create order from cart
-    const order = await Order.create({
-      ...req.body,
-      status: req.body.payMethod === "cod" ? "placed" : "pending",
-      userId: req.user._id,
-    });
-    await Cart.deleteOne({ userId: req.user._id });
-    // decrease product stock after order is placed successfully
-    const products = await Product.find({
-      _id: { $in: order.products },
-    });
-    products.forEach(async (product) => {
-      product.stock = product.stock - 1;
-      await product.save();
-    });
 
-    res.status(200).json({
-      message: "Order created successfully",
-      order,
-      success: true,
-    });
+    if (req.body.payMethod === "cod") {
+      let order = await Order.create({
+        ...req.body,
+        status: req.body.payMethod === "cod" ? "cod" : "pending",
+        userId: req.user._id,
+      });
+      //  update cart
+      await Cart.deleteOne({ userId: req.user._id });
+      // substract product stock after order is placed successfully
+      const products = await Product.find({
+        _id: { $in: order.products },
+      });
+      products.forEach(async (product) => {
+        await Product.findByIdAndUpdate(product._id, {
+          stock:
+            product.stock -
+            order.products.find(
+              (p) => p.productId.toString() === product._id.toString()
+            ).quantity,
+        });
+      });
+      res.status(200).json({
+        message: "Order created successfully",
+        order,
+        success: true,
+      });
+    } else {
+      let order = await Order.create({
+        ...req.body,
+        status: "pending",
+        userId: req.user._id,
+        razorpayOrderId: req.body.response.razorpay_order_id,
+        razorpayPaymentId: req.body.response.razorpay_payment_id,
+        razorpaySignature: req.body.response.razorpay_signature,
+      });
+      res.status(200).json({
+        message: "Order created successfully",
+        order,
+        success: true,
+      });
+    }
   } catch (error) {
     console.log(error);
     res.status(400).json({
@@ -49,11 +81,64 @@ exports.createOrder = async (req, res) => {
   }
 };
 
+exports.verifyPayment = async (req, res) => {
+  try {
+    let body = req.body.razorpayOrderId + "|" + req.body.razorpayPaymentId;
+    let key = process.env.RAZORPAY_SECRET;
+    var crypto = require("crypto");
+    var expectedSignature = crypto
+      .createHmac("sha256", key)
+      .update(body.toString())
+      .digest("hex");
+    var response = { signatureIsValid: "false" };
+    if (expectedSignature === req.body.razorpaySignature) {
+      response = { signatureIsValid: "true" };
+      let order = await Order.findOne({
+        razorpayOrderId: req.body.razorpayOrderId,
+      });
+      if (order.status === "pending") {
+        order.status = "paid";
+        order.save();
+      }
+      await Cart.deleteOne({ userId: req.user._id });
+      // decrease product stock after order is placed successfully
+      const products = await Product.find({
+        _id: { $in: order.products },
+      });
+      products.forEach(async (product) => {
+        await Product.findByIdAndUpdate(product._id, {
+          stock:
+            product.stock -
+            order.products.find(
+              (p) => p.productId.toString() === product._id.toString()
+            ).quantity,
+        });
+      });
+      res.status(200).json({
+        message: "Payment verified successfully",
+        response,
+        success: true,
+      });
+    } else {
+      res.status(400).json({
+        message: "Payment verification failed",
+        response,
+        success: false,
+      });
+    }
+  } catch (error) {
+    console.log(error);
+    res.status(400).json({
+      error,
+    });
+  }
+};
 // get my orders
 exports.getMyOrders = async (req, res) => {
   try {
     const orders = await Order.aggregate([
       { $match: { userId: req.user._id } },
+      { $match: { status: { $ne: "pending" } } },
       { $unwind: "$products" },
       {
         $lookup: {
@@ -76,16 +161,25 @@ exports.getMyOrders = async (req, res) => {
   }
 };
 
-exports.placeOrder = async (req, res) => {
+exports.createOrderId = async (req, res) => {
   try {
-    const order = await Order.findById(req.params.id);
-    order.status = "placed";
-    const updatedOrder = await order.save();
-    res.status(200).json({
-      message: "Order placed successfully",
-      updatedOrder,
+    const options = {
+      amount: req.body.amount * 100,
+      currency: "INR",
+      receipt: "receipt#1",
+      payment_capture: "1",
+    };
+    instance.orders.create(options, (err, order) => {
+      if (err) {
+        console.log(err);
+      } else {
+        res.status(200).json({
+          order,
+        });
+      }
     });
   } catch (error) {
+    console.log(error);
     res.status(400).json({
       error,
     });
